refactor(core): extract search module logic to core-lib crate
Move serialization, index management, and search_keys logic from
napi and wasm binding crates to core-lib. Eliminates ~700 lines of
duplicated code between the two binding layers.

Closes #509
